Update AzureFunction.v1.csx
If someone has created Azure function with run-time version 2 - then in JSON response the keys gets converted into small case. Due to this the Azure DevOps dashboard - Twitter sentiment widget was not working and showing NaN values. 

To fix the above issue with Azure function run-time version 2- Added JsonProperty attribute for the variables. So that the JSON response would be return as Widget expected.